Update README.md
Added Process and Methodology section to README

This addition to the README expands on the methodology of the project and gives a general overview of each step in the process, plus its code location. Furthermore, this will make it easier for future contributors to make changes and expand on the code based on the future work and model improvement section, as they will have a better understanding of the file structure and training pipeline.